refactor module, refactor Dockerfile and webserver config, update README
