+++ backend/api/wsgi.py
@@ -1,4 +1,4 @@
-from app.main import app
+from api.app import app
 
 if __name__ == "__main__":
     app.run()
