+++ backend/webserver-config/uwsgi.ini
@@ -1,5 +1,5 @@
 [uwsgi]
-module = app.main
+module = api.app
 callable = app
 
 uid = nginx
